update storage limit check logic

+++ ProtonMail/ProtonMail/AppConstants.swift
@@ -13,7 +13,7 @@ class AppConstants {
     
     static let CacheVersion : Int = 98
     static let AuthCacheVersion : Int = 15
-    static let SpaceWarningThreshold: Double = 80
+    static let SpaceWarningThreshold: Int = 80
     static let SplashVersion : Int = 1
     static let TourVersion : Int = 2
